Refactor autoencoder training and inference
- Replace CelebaDataLoader with generic ImageDataLoader for better reusability
- Update autoencoder training and inference scripts to use new dataloader
- Remove CelebaDataLoader dependency and clean up imports
- Update README with improved documentation and usage examples
- Fix test imports to use new dataloader structure
